support deb packaging on ubuntu trusty

+++ Makefile
@@ -15,7 +15,10 @@ install:
 	cp $(src_dir)/$(PROJECT) $(DESTDIR)/usr/bin
 	mkdir -p $(DESTDIR)/etc/kafka2influxdb
 	cp $(src_dir)/kafka2influxdb.example.toml $(DESTDIR)/etc/kafka2influxdb/kafka2influxdb.toml
-
+	if [ -f /usr/lib/upstart ]; then
+		mkdir -p $(DESTDIR)/etc/init
+		echo "manual" >> /etc/init/kafka2influxdb.override
+	fi
 clean:
 	rm -f $(src_dir)/$(PROJECT)
 	rm -rf /tmp/kafka2influxdb
@@ -24,17 +27,31 @@ distclean: clean
 
 deb:
 	if ! dpkg -s devscripts > /dev/null 2>&1 ; then sudo apt-get install devscripts; fi
-	if ! dpkg -s dh-systemd > /dev/null 2>&1 ; then sudo apt-get install dh-systemd; fi
 	cp -f $(src_dir)/kafka2influxdb.1 $(src_dir)/debian
-	cp -f $(src_dir)/kafka2influxdb.service $(src_dir)/debian
-	sed -i 's/local\///' -i debian/kafka2influxdb.service
+	if [ -f /usr/lib/upstart ]; then
+		cp -f $(src_dir)/kafka2influxdb.upstart $(src_dir)/debian
+		cp -f $(src_dir)/debian/rules.nonsystemd $(src_dir)/debian/rules
+		sed -i 's/local\///' -i debian/kafka2influxdb.upstart
+		.
+	elif [ -f /lib/systemd/systemd ]; then
+		if ! dpkg -s dh-systemd > /dev/null 2>&1 ; then sudo apt-get install dh-systemd; fi
+		cp -f $(src_dir)/kafka2influxdb.service $(src_dir)/debian
+		cp -f $(src_dir)/debian/rules.systemd $(src_dir)/debian/rules
+		sed -i 's/local\///' -i debian/kafka2influxdb.service
+		.
+	else
+		cp -f $(src_dir)/kafka2influxdb.init $(src_dir)/debian
+		cp -f $(src_dir)/debian/rules.nonsystemd $(src_dir)/debian/rules
+		sed -i 's/local\///' -i debian/kafka2influxdb.init
+		.
+	fi
 	dpkg-buildpackage -us -uc -b
 
 manpage:
 	pandoc $(src_dir)/manpage.md -s -t man -o $(src_dir)/kafka2influxdb.1
 
 bindata: manpage
-	go-bindata $(src_dir)/kafka2influxdb.service $(src_dir)/kafka2influxdb.1
+	go-bindata kafka2influxdb.service kafka2influxdb.1
 
 sync: bindata
 	git add . && git commit && git pull origin master && git push origin master
